refactor: Card pool 코드 분리
