+++ README.md
@@ -103,6 +103,6 @@
 - [x] enum 활용해 if문 제거
 - [ ] 함수형 인터페이스 감싸서 클래스에 의미 부여
 - [x] stream map 내부 출력문 변경
-- [ ] Card pool 코드 수정
+- [x] Card pool 코드 분리
 - [x] 딜러의 손패를 보여주는 행동을 전략패턴을 이용하도록 변경
 - [ ] 플레이어와 딜러에서 카드를 선택하는 전략을 선택할 수 없도록 하기
